LayoutTests/editing/smart-lists/ordered-and-unordered-mismatch-does-not-trigger-conversion-expected.txt

An ordered marker followed by an unordered marker (or vice versa) on consecutive lines should not trigger smart-list conversion, because ordered and unordered markers are incompatible.

Unordered then ordered:
| "* Hello"
| <div>
|   "1. World<#selection-caret>"

Ordered then unordered:
| "1. Hello"
| <div>
|   "* World<#selection-caret>"
